Changed hotkey to Mute and holding Home exits ES
Changed hotkey to be Mute instead of Select because some games make use of the Select button, made changes in retrogame.c to allow Home to be used to exit games even though it's wired to be Start and Select combo, rebuilt retrogame, and changed retrogame.cfg to allow use of Mute button and holding Home now triggers F4 button, useful when without a keyboard to get to command line. With PSPI F4, you can now exit to terminal, arrow up or down to find commands, and press X to execute. I use it to run setup.sh when I don't have a keyboard close.